+++ src/Oro/Bundle/FormBundle/Tests/Unit/Form/Extension/TooltipFormExtensionTest.php
@@ -117,7 +117,7 @@ class TooltipFormExtensionTest extends \PHPUnit\Framework\TestCase
         $this->assertEquals($tooltip, $view->vars['tooltip']);
     }
 
-    public function testModifyEntityConfigTooltipFieldWithMessagesTranslationDomain()
+    public function testModifyEntityConfigTooltipFieldWithTooltipTranslation()
     {
         $tooltip = 'Custom Tooltip';
 
@@ -153,56 +153,8 @@ class TooltipFormExtensionTest extends \PHPUnit\Framework\TestCase
 
         $this->translator->expects($this->once())
             ->method('hasTrans')
-            ->with($tooltip, 'messages')
+            ->with($tooltip)
             ->willReturn(true);
-
-        $view = new FormView();
-        $this->extension->buildView($view, $form, []);
-
-        $this->assertArrayHasKey('tooltip', $view->vars);
-        $this->assertEquals($tooltip, $view->vars['tooltip']);
-    }
-
-    public function testModifyEntityConfigTooltipFieldWithTooltipTranslationDomain()
-    {
-        $tooltip = 'Custom Tooltip';
-
-        $form = $this->createMock(FormInterface::class);
-        $parentForm = $this->createMock(FormInterface::class);
-        $parentFormConfig = $this->createMock(FormConfigInterface::class);
-        $form->expects($this->any())
-            ->method('getParent')
-            ->willReturn($parentForm);
-        $form->expects($this->any())
-            ->method('getName')
-            ->willReturn('testField');
-        $parentForm->expects($this->once())
-            ->method('getConfig')
-            ->willReturn($parentFormConfig);
-        $parentFormConfig->expects($this->once())
-            ->method('getOptions')
-            ->willReturn(['data_class' => 'TestClass']);
-
-        $config = $this->createMock(ConfigInterface::class);
-        $this->configProvider->expects($this->any())
-            ->method('hasConfig')
-            ->with('TestClass', 'testField')
-            ->willReturn(true);
-        $this->configProvider->expects($this->once())
-            ->method('getConfig')
-            ->with('TestClass', 'testField')
-            ->willReturn($config);
-        $config->expects($this->once())
-            ->method('get')
-            ->with('description')
-            ->willReturn($tooltip);
-
-        $this->translator->expects($this->exactly(2))
-            ->method('hasTrans')
-            ->willReturnMap([
-                [$tooltip, 'messages', null, false],
-                [$tooltip, 'tooltips', null, true]
-            ]);
 
         $view = new FormView();
         $this->extension->buildView($view, $form, []);