+++ CHANGELOG.md
@@ -9,8 +9,18 @@ The current file describes significant changes in the code that may affect the u
 
 ### Removed
 
+#### CronBundle
+* `Oro\Bundle\CronBundle\Command\CronCommandInterface` has been removed.
+  Use `Oro\Bundle\CronBundle\Command\CronCommandScheduleDefinitionInterface`
+  and `Oro\Bundle\CronBundle\Command\CronCommandActivationInterface` instead.
+
+#### DataGridBundle
+* The deprecated `options / skip_acl_check` datagrid option was removed. Use the `source / skip_acl_apply` option instead. 
+* The deprecated `source / acl_resource` datagrid option was removed. Use the `acl_resource` option instead.
+
 #### UIBundle
-* `oroui/js/app/views/input-widget/checkbox`was removed; use pure CSS checkbox customization instead.
+* `oroui/js/app/views/input-widget/checkbox` was removed; use pure CSS checkbox customization instead.
+* The deprecated `tooltips` translation domain was removed. All translation from this domain were moved to the `messages` domain.
 
 ## 5.1.0-alpha.1 (2022-05-31)
 [Show detailed list of changes](incompatibilities-5-1-alpha.md)